Move customized Flask class to it's own file
