+++ api/yangCatalogApi.py
@@ -40,36 +40,19 @@ __license__ = 'Apache License, Version 2.0'
 __email__ = 'miroslav.kovac@pantheon.tech'
 
 import collections
-import configparser
-import errno
-import grp
 import json
-import logging
 import os
-import pwd
-import shutil
 import sys
-import threading
 import time
 import uuid
-from datetime import datetime, timedelta
 from threading import Lock
 
-import requests
-import flask
-from elasticsearch import Elasticsearch
-from flask import (Config, Flask, Response, abort, jsonify, make_response,
-                   redirect, request)
-from flask.logging import default_handler
+from flask import abort, g, jsonify, make_response, redirect, request
+from flask.wrappers import Response
 from flask_cors import CORS
-from flask_pyoidc.user_session import UserSession
-from redis import Redis
-from redisConnections.redisConnection import RedisConnection
-from utility.confdService import ConfdService
-from utility.redisUsersConnection import RedisUsersConnection
 
 import api.authentication.auth as auth
-from api.sender import Sender
+from api.my_flask import MyFlask
 from api.views.admin.admin import bp as admin_bp
 from api.views.admin.admin import ietf_auth
 from api.views.errorHandlers.errorHandler import bp as error_handling_bp
@@ -80,354 +63,6 @@ from api.views.yangSearch.yangSearch import bp as yang_search_bp
 from api.views.ycJobs.ycJobs import bp as jobs_bp
 from api.views.ycSearch.ycSearch import bp as search_bp
 
-
-class MyConfig(Config):
-
-    def __getattr__(self, name: str):
-        try:
-            return self[name.upper().replace('_', '-')]
-        except KeyError:
-            raise AttributeError(name)
-
-
-class MyFlask(Flask):
-
-    config_class = MyConfig
-
-    def __init__(self, import_name):
-        self.loading = True
-        super(MyFlask, self).__init__(import_name)
-        self.ys_set = 'set'
-        self.waiting_for_reload = False
-        self.response_waiting = None
-        self.special_id = None
-        self.special_id_counter = {}
-        self.release_locked = []
-        self.permanent_session_lifetime = timedelta(minutes=20)
-        self.load_config()
-        self.logger.debug('API initialized at {}'.format(self.config.yangcatalog_api_prefix))
-        self.logger.debug('Starting api')
-        self.secret_key = self.config.s_flask_secret_key
-        self.confdService = ConfdService()
-        self.redisConnection = RedisConnection()
-
-    def load_config(self):
-        self.init_config()
-        self.config.from_file(os.environ['YANGCATALOG_CONFIG_PATH'], load=self.config_reader)
-        self.setup_logger()
-        self.post_config_load()
-
-    def config_reader(self, file):
-        parser = configparser.ConfigParser()
-        parser._interpolation = configparser.ExtendedInterpolation()
-        parser.read(file.name)
-        mapping = {}
-        for section in parser.sections():
-            section_prefix = ''.join((x for x in section.split('-')[0] if x.isupper()))
-            for key, value in parser.items(section):
-                key = '{}-{}'.format(section_prefix, key.upper())
-                mapping[key] = value
-        mapping['CONFIG-PARSER'] = parser
-        return mapping
-
-    def init_config(self):
-        self.config['LOCK-UWSGI-CACHE1'] = threading.Lock()
-        self.config['LOCK-UWSGI-CACHE2'] = threading.Lock()
-
-    def setup_logger(self):
-        self.logger.removeHandler(default_handler)
-        file_name_path = '{}/yang.log'.format(self.config.d_logs)
-        os.makedirs(os.path.dirname(file_name_path), exist_ok=True)
-        exists = False
-        if os.path.isfile(file_name_path):
-            exists = True
-        FORMAT = '%(asctime)-15s %(levelname)-8s %(filename)s api => %(message)s - %(lineno)d'
-        DATEFMT = '%Y-%m-%d %H:%M:%S'
-        handler = logging.FileHandler(file_name_path)
-        handler.setFormatter(logging.Formatter(FORMAT, DATEFMT))
-        self.logger.setLevel(logging.DEBUG)
-        self.logger.addHandler(handler)
-        if not exists:
-            os.chmod(file_name_path, 0o664)
-
-    def post_config_load(self):
-        self.config['S-ELK-CREDENTIALS'] = self.config.s_elk_secret.strip('"').split()
-        self.config['S-CONFD-CREDENTIALS'] = self.config.s_confd_credentials.strip('"').split()
-        self.config['DB-ES-AWS'] = self.config.db_es_aws == 'True'
-        if self.config.db_es_aws:
-            self.config['ES'] = Elasticsearch([self.config.db_es_host],
-                                              http_auth=(self.config.s_elk_credentials[0],
-                                                         self.config.s_elk_credentials[1]),
-                                              scheme='https', port=443)
-        else:
-            self.config['ES'] = Elasticsearch([{'host': '{}'.format(self.config.db_es_host), 'port': self.config.db_es_port}])
-
-        rabbitmq_host = self.config.config_parser.get('RabbitMQ-Section', 'host', fallback='127.0.0.1')
-        rabbitmq_port = int(self.config.config_parser.get('RabbitMQ-Section', 'port', fallback='5672'))
-        rabbitmq_virtual_host = self.config.config_parser.get('RabbitMQ-Section', 'virtual-host', fallback='/')
-        rabbitmq_username = self.config.config_parser.get('RabbitMQ-Section', 'username', fallback='guest')
-        rabbitmq_password = self.config.config_parser.get('Secrets-Section', 'rabbitMq-password', fallback='guest')
-        self.config['SENDER'] = Sender(
-            self.config.d_logs, self.config.d_temp,
-            rabbitmq_host=rabbitmq_host,
-            rabbitmq_port=rabbitmq_port,
-            rabbitmq_virtual_host=rabbitmq_virtual_host,
-            rabbitmq_username=rabbitmq_username,
-            rabbitmq_password=rabbitmq_password
-        )
-
-        separator = ':'
-        suffix = self.config.w_api_port
-        if self.config.g_uwsgi == 'True':
-            separator = '/'
-            suffix = 'api'
-        self.config['YANGCATALOG-API-PREFIX'] = '{}://{}{}{}/' \
-            .format(self.config.g_protocol_api, self.config.w_ip, separator, suffix)
-
-        self.config.g_is_prod = self.config.g_is_prod == 'True'
-        self.config['REDIS'] = Redis(
-            host=self.config.db_redis_host,
-            port=self.config.db_redis_port
-        )
-        self.config['REDIS-USERS'] = RedisUsersConnection()
-        auth.users = self.config.redis_users
-        self.check_wait_redis_connected()
-
-    def check_wait_redis_connected(self):
-        while not self.config.redis.ping():
-            time.sleep(5)
-            self.logger.info('Waiting 5 seconds for redis to start')
-
-    def process_response(self, response):
-        response = super().process_response(response)
-        self.create_response_only_latest_revision(response)
-        # self.create_response_with_yangsuite_link(response)
-
-        try:
-            if request.special_id != 0:
-                if request.special_id not in self.release_locked:
-                    self.release_locked.append(request.special_id)
-                    self.response_waiting = response
-                    self.waiting_for_reload = False
-        except:
-            pass
-        return response
-
-    def preprocess_request(self):
-        super().preprocess_request()
-        request.special_id = 0
-        if not 'admin' in request.path:
-            app.logger.info(request.path)
-        if 'api/admin' in request.path and not 'api/admin/healthcheck' in request.path and not 'api/admin/ping' in request.path:
-            logged_in = UserSession(flask.session, 'default').is_authenticated()
-            app.logger.info('User logged in {}'.format(logged_in))
-            if self.config.g_is_prod and not logged_in and 'login' not in request.path:
-                return abort(401, description='not yet Authorized')
-
-    def create_response_only_latest_revision(self, response):
-        if request.args.get('latest-revision'):
-            if 'True' == request.args.get('latest-revision'):
-                if response.data:
-                    decoded_string = response.data.decode(encoding='utf-8', errors='strict')
-                    json_data = json.JSONDecoder(object_pairs_hook=collections.OrderedDict).decode(decoded_string)
-                else:
-                    return response
-                modules = None
-                if json_data.get('yang-catalog:modules') is not None:
-                    if json_data.get('yang-catalog:modules').get(
-                            'module') is not None:
-                        modules = json_data.get('yang-catalog:modules').get(
-                            'module')
-                elif json_data.get('module') is not None:
-                    modules = json_data.get('module')
-                modules_to_remove = []
-                if modules:
-                    if len(modules) > 0:
-                        newlist = sorted(modules, key=lambda k: k['name'])
-                        temp_module = {}
-                        i = 0
-                        for mod in newlist:
-                            name = mod['name']
-                            if temp_module:
-                                if temp_module['name'] == name:
-                                    revisions = []
-                                    mod['index'] = i
-                                    year = int(temp_module['revision'].split('-')[0])
-                                    month = int(temp_module['revision'].split('-')[1])
-                                    day = int(temp_module['revision'].split('-')[2])
-                                    try:
-                                        revisions.append(datetime(year, month, day))
-                                    except ValueError:
-                                        if day == 29 and month == 2:
-                                            revisions.append(datetime(year, month, 28))
-                                    year = int(mod['revision'].split('-')[0])
-                                    month = int(mod['revision'].split('-')[1])
-                                    day = int(mod['revision'].split('-')[2])
-                                    try:
-                                        revisions.append(datetime(year, month, day))
-                                    except ValueError:
-                                        if day == 29 and month == 2:
-                                            revisions.append(datetime(year, month, 28))
-                                    latest = revisions.index(max(revisions))
-                                    if latest == 0:
-                                        modules_to_remove.append(mod['index'])
-                                    elif latest == 1:
-                                        modules_to_remove.append(temp_module['index'])
-                                else:
-                                    mod['index'] = i
-                                    temp_module = mod
-                            else:
-                                mod['index'] = i
-                                temp_module = mod
-                            i += 1
-                        for mod_to_remove in reversed(modules_to_remove):
-                            newlist.remove(newlist[mod_to_remove])
-                        for mod in newlist:
-                            if mod.get('index'):
-                                del mod['index']
-                        response.data = json.dumps(newlist)
-
-    def get_dependencies(self, mod, mods, inset):
-        if mod.get('dependencies'):
-            for dep in mod['dependencies']:
-                if dep['name'] in inset:
-                    continue
-                if dep.get('revision'):
-                    mods.add(dep['name'] + '@' + dep[
-                        'revision'] + '.yang')
-                    inset.add(dep['name'])
-                    search_filter = json.dumps({
-                        'input': {
-                            'name': dep['name'],
-                            'revision': dep['revision']
-                        }
-                    })
-                    rp = requests.post('{}search-filter'.format(
-                        self.config.yangcatalog_api_prefix), search_filter,
-                        headers={
-                            'Content-type': 'application/json',
-                            'Accept': 'application/json'}
-                    )
-                    mo = rp.json()['yang-catalog:modules']['module'][0]
-                    self.get_dependencies(mo, mods, inset)
-                else:
-                    rp = requests.get('{}search/name/{}'
-                                      .format(self.config.yangcatalog_api_prefix,
-                                              dep['name']))
-                    if rp.status_code == 404:
-                        continue
-                    mo = rp.json()['yang-catalog:modules']['module']
-                    revisions = []
-                    for m in mo:
-                        revision = m['revision']
-                        year = int(revision.split('-')[0])
-                        month = int(revision.split('-')[1])
-                        day = int(revision.split('-')[2])
-                        revisions.append(datetime(year, month, day))
-                    latest = revisions.index(max(revisions))
-                    inset.add(dep['name'])
-                    mods.add('{}@{}.yang'.format(dep['name'],
-                                                 mo[latest][
-                                                     'revision']))
-                    self.get_dependencies(mo[latest], mods, inset)
-
-    def create_response_with_yangsuite_link(self, response):
-        if request.headers.environ.get('HTTP_YANGSUITE'):
-            if 'true' != request.headers.environ['HTTP_YANGSUITE']:
-                return response
-            if request.headers.environ.get('HTTP_YANGSET_NAME'):
-                self.ys_set = request.headers.environ.get(
-                    'HTTP_YANGSET_NAME').replace('/', '_')
-        else:
-            return response
-
-        if response.data:
-            json_data = json.loads(response.data)
-        else:
-            return response
-        modules = None
-        if json_data.get('yang-catalog:modules') is not None:
-            if json_data.get('yang-catalog:modules').get('module') is not None:
-                modules = json_data.get('yang-catalog:modules').get('module')
-        elif json_data.get('module') is not None:
-            modules = json_data.get('module')
-        if modules:
-            if len(modules) > 0:
-                ys_dir = self.config.d_ys_users
-
-                id = uuid.uuid4().hex
-                ys_dir += '/' + id + '/repositories/' + modules[0]['name'].lower()
-                try:
-                    os.makedirs(ys_dir)
-                except OSError as e:
-                    # be happy if someone already created the path
-                    if e.errno != errno.EEXIST:
-                        return 'Server error - could not create directory'
-                mods = set()
-                inset = set()
-                if len(modules) == 1:
-                    defmod = modules[0]['name']
-                for mod in modules:
-                    name = mod['name']
-                    if name in inset:
-                        continue
-                    name += '@{}.yang'.format(mod['revision'])
-                    mods.add(name)
-                    inset.add(mod['name'])
-                    self.get_dependencies(mod, mods, inset)
-                if (('openconfig-interfaces' in inset
-                     or 'ietf-interfaces' in inset)
-                        and 'iana-if-type' not in inset):
-                    resp = requests.get(
-                        '{}search/name/iana-if-type?latest-revision=True'.format(self.config.yangcatalog_api_prefix),
-                        headers={
-                            'Content-type': 'application/json',
-                            'Accept': 'application/json'})
-                    name = '{}@{}.yang'.format(resp.json()[0]['name'],
-                                               resp.json()[0]['revision'])
-                    mods.add(name)
-                    inset.add(resp.json()[0]['name'])
-                    self.get_dependencies(resp.json()[0], mods, inset)
-
-                modules = []
-                for mod in mods:
-                    if not os.path.exists(self.config.d_save_file_dir + '/' + mod):
-                        continue
-                    shutil.copy(self.config.d_save_file_dir + '/' + mod, ys_dir)
-                    modules.append([mod.split('@')[0], mod.split('@')[1].replace('.yang', '')])
-                ys_dir = self.config.d_ys_users
-                ys_dir += '/' + id + '/yangsets'
-                try:
-                    os.makedirs(ys_dir)
-                except OSError as e:
-                    # be happy if someone already created the path
-                    if e.errno != errno.EEXIST:
-                        return 'Server error - could not create directory'
-                json_set = {'owner': id,
-                            'repository': id + '+' + defmod.split('@')[0],
-                            'setname': defmod.split('@')[0],
-                            'defmod': defmod.split('@')[0],
-                            'modules': modules}
-
-                with open(ys_dir + '/' + defmod.split('@')[0].lower(), 'w') as f:
-                    f.write(json.dumps(json_set, indent=4))
-                uid = pwd.getpwnam('yang').pw_uid
-                gid = grp.getgrnam('yang-dev').gr_gid
-                path = self.config.d_ys_users + '/' + id
-                for root, dirs, files in os.walk(path):
-                    for momo in dirs:
-                        os.chown(os.path.join(root, momo), uid, gid)
-                    for momo in files:
-                        os.chown(os.path.join(root, momo), uid, gid)
-                os.chown(path, uid, gid)
-                json_data['yangsuite-url'] = (
-                    '{}yangsuite/{}'.format(self.config.yangcatalog_api_prefix, id))
-                response.data = json.dumps(json_data)
-                return response
-            else:
-                return response
-        else:
-            return response
 
 
 app = MyFlask(__name__)
@@ -558,7 +193,7 @@ def load():
             app.logger.info('Application locked for reload')
             app.waiting_for_reload = True
             special_id = str(uuid.uuid4())
-            request.special_id = special_id
+            g.special_id = special_id
             app.special_id = special_id
             app.special_id_counter[special_id] = 0
             app.logger.info('Special ids {}'.format(app.special_id_counter))
